ref: Move OrganizationDetails views into correct folder
Move the organization details view into a folder, move the install banner
into views as well since it's not a resuable component.
