+++ src/sentry/static/sentry/app/views/organizationDetails/installPromptBanner.jsx
