+++ tests/js/spec/views/organizationDetails/organizationsDetails.spec.jsx
